+++ design-system/popper/src/popper.tsx
@@ -1,4 +1,4 @@
-import React, { Fragment, useCallback, useMemo } from 'react';
+import React, { useMemo } from 'react';
 
 import { Placement, VirtualElement } from '@popperjs/core';
 import {
@@ -7,8 +7,6 @@ import {
   PopperProps,
   Popper as ReactPopper,
 } from 'react-popper';
-
-import { getBooleanFF } from '@atlaskit/platform-feature-flags';
 
 export { placements } from '@popperjs/core';
 // Export types from PopperJS / React Popper
@@ -87,11 +85,6 @@ function defaultChildrenFn() {
 
 const defaultOffset: Offset = [0, 8];
 
-let siblingElement: JSX.Element | null = null;
-if (getBooleanFF('uip.popper.flex-team')) {
-  siblingElement = <em hidden data-flex="Hello from UIP"></em>;
-}
-
 export function Popper<CustomModifiers>({
   children = defaultChildrenFn,
   offset = defaultOffset,
@@ -126,18 +119,6 @@ export function Popper<CustomModifiers>({
     return [...internalModifiers, ...modifiers];
   }, [internalModifiers, modifiers]);
 
-  const wrappedChildren = useCallback(
-    (childrenProps) => {
-      return (
-        <Fragment>
-          {siblingElement}
-          {children(childrenProps)}
-        </Fragment>
-      );
-    },
-    [children],
-  );
-
   return (
     <ReactPopper
       modifiers={mergedModifiers}
@@ -145,7 +126,7 @@ export function Popper<CustomModifiers>({
       strategy={strategy}
       referenceElement={referenceElement}
     >
-      {wrappedChildren}
+      {children}
     </ReactPopper>
   );
 }