+++ design-system/tokens/scripts/style-dictionary/constants.tsx
@@ -1,3 +1,3 @@
 export const THEME_INPUT_DIR = './src/tokens/';
 export const ARTIFACT_OUTPUT_DIR = './src/artifacts/';
-export const FIGMA_ARTIFACT_OUTPUT_DIR = './src/figma/';
+export const FIGMA_ARTIFACT_OUTPUT_DIR = './figma/';