+++ design-system/tokens/src/entry-points/figma-sync.tsx
@@ -1,4 +1,0 @@
-import atlassianDark from '../figma/atlassian-dark.json';
-import atlassianLight from '../figma/atlassian-light.json';
-
-export { atlassianLight, atlassianDark };